Updated scripts to reset folders in benchmark mode and to loop across 5 simulation steps

+++ libraries/miscellaneous/empty_results.sh
@@ -1,6 +1,10 @@
 # Script for deleting existing SCORE results
 # This is for testing various experimental combinations and parameters
 # FastQC and Flexbar files should not need to be deleted
+# To empty all results folders: ./empty_results.sh full
+# To empty all results folders and save previous results: ./empty_results.sh full Test_Folder
+
+MODE=$1
 
 cd simulation_data/
 rm -r *.gz
@@ -8,19 +12,42 @@ rm -r *.idx
 rm -r kallisto/sample*
 rm -r *.rda
 rm -r *.txt
+
 cd ../../../deg/
-rm -r consensus*
-rm -r deg*
-rm -r diffexpr*
-rm -r filtered*
-rm -r pathway*
-#cd ../fastqc/
-#rm -r *
+
+if [ -z "$2" ]
+then
+	rm -r consensus*
+	rm -r deg*
+	rm -r diffexpr*
+	rm -r filtered*
+	rm -r pathway*
+else
+	ANALYSIS=$2
+	mkdir $ANALYSIS
+	mv consensus* $ANALYSIS
+	mv deg* $ANALYSIS
+	mv diffexpr* $ANALYSIS
+	mv filtered* $ANALYSIS
+	mv pathway* $ANALYSIS
+fi
+
+if [ "$MODE" = "full" ]
+then
+	cd ../fastqc/
+	rm -r *
+fi
+
 cd ../mapped/bowtie2/featureCounts/
 rm -r *
 cd ../../kallisto/
 rm -r *
 cd ../../trimmed/
-#rm -r *
+
+if [ "$MODE" = "full" ]
+then
+	rm -r *
+fi
+
 cd ../
 rm log.txt
